tooling/templatize: load settings file everywhere
While we can continue exposing the raw flags for setting cloud,
environment, region, stamp, etc, it's most likely that users will prefer
to use the settings file and specify just the specific environment they
are interested in.

Now that the service configuration validator knows how to read and
evaluate every setting, we can remove the old materialize target and
JSON files.

We are getting close to removing the templatize.sh script entirely,
which is great - that means that the CLI itself is easy to use and it's
now portable to other projects that want it.

Signed-off-by: Steve Kuznetsov <stekuznetsov@microsoft.com>