make RP image digests explicit in dev config
in dev environments, the rp image digests were undefined and we had a deployment mechanism in place to lookup digests at deploytime. this was supposed to bring some dev-ease-of-life. but it also makes reasoning about the config harder and is problematic for processes that require proper configuration.

this PR proposed the following changes:
* RH config needs to mention the image digests and the `make deploy` targets as well as the GH action based delivery will rely on that
* there will be an additional target `make deploy.current` that will provide the current behaviour and can be used in personal dev environments

this implies that 1P needs to bump the digests in config at relevant points in time to make sure the changes that should go to integrated dev are actually promoted.

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>