tooling/templatize: manage working directory in-memory
Mutating global state (the current directory for the process) is
generally a bad idea - not only does this make it impossible to run
multiple pipelines in parallel but it does so by creating implicit
raciness hidden by the OS interface and not visible to e.g. the `-race`
checker in Go tests.

Signed-off-by: Steve Kuznetsov <stekuznetsov@microsoft.com>