skip container instance on RG deletion
the deletion of an RG is executed in a container instance in the same RG. so we need to skip container instances in order to now kill the deletion step.
since we don't use container instances for anything else, we don't leave anything blocking behind by this. ev2 cleans up the container instance after it finished
Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>