frontend: Delete stale Cosmos DB data upon detection
If Cluster Service returns "404 Not Found" to a cluster or node
pool deletion request, that means we have stale data in Cosmos DB.

This has been observed in test environments where a user bypasses
the RP to delete a resource (e.g. "ocm delete").

This can also occur if an asychronous deletion operation fails,
which was observed recently in a CI job run in the INT environment.

To date, the only way to recover from this was to manually delete
the stale data from Cosmos DB. This change makes the RP frontend
proactively delete the stale Cosmos DB data upon detection.

There is still an open question about how to correctly handle a
failed asynchronous deletion. That will be addressed separately.