compile bicep into json for e2e test setup option
This is a basic update/verify pattern to create compiled bicep templates
for setup of zero-dep e2e tests to be created during the testing.
We use hashes of the input and the output to decide whether to
regenerate.
We will not call the verify script because we will not commit the
generated files to avoid conflicts on minor versions of az.