*: use input/output chaining for the global MSI (#2185)
Due to a regretful oversight, the full resource ID of the global MSI was
put into the configuration. However, this means that the global pipeline
is no longer self-contained and rollouts to new tenants cannot be done
hermetically, as after the subscription is provisioned, we would need to
update the config to refer back to that ID. Instead, we can just use an
input/output chain to look this value up at runtime.

Signed-off-by: Steve Kuznetsov <stekuznetsov@microsoft.com>