move image digest configuration for MSFT environments back to ARO-HCP (#2182)
* move image digest configuration for MSFT environments back to ARO-HCP

with we moved the msft configuration overlay into the ADO sdp-pipelines repository.
this introduced that drawback, that we lost the ability to reason about image digests across environments.

this PR brings these image digests back to the ARO-HCP repo in the form of the config/config.msft.clouds-overlay.yaml file. the promotion process of MSFT image bumps starts with editing this file. the sdp-pipelines update process will consume this overlay file and merge it into the final configuration for MSFT.

if required, we can move more configuration non-sensitive properties back from sdp-pipelines to the ARO-HCP repo in the future.

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>

* bump rp backend/frontend

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>

---------

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>