update CS image to include MIs and MI DP interaction fix in integrated dev (#2265)
* update CS image to include MIs and MI DP interaction fix in integrated dev

We update the CS image to the
8b2e9af554ad3422f5c30830e383eddffdeac37f64677359d6913f53a4a2abfb
SHA. This SHA is associated to the f973dd9 CS commit.

* update CS image to include MIs and MI DP interaction fix in int, stage, prod

We update the CS image to the
8b2e9af554ad3422f5c30830e383eddffdeac37f64677359d6913f53a4a2abfb
SHA. This SHA is associated to the f973dd9 CS commit.

We also remove the image registry override in stage as it is not
needed anymore with the MIs and MI DP interaction fix.