optionally generate dedicated CRD chart from OLM bundle
`olm-bundle-repkg` usually places the CRDs of an OLM bundle into the `crd` folders of the chart. but these CRDs are only applied once during first install and are never touched again. this is by design of helm. they want to avoid that people loose their "data" by reinstalling an operator.

proposal: we introduce the option the `olm-bundle-rpkg` to place CRDs into their own chart into the `templates` folder. manifests from the `templates` folder are applied to the cluster even on updates.

with this we gain the ability to update CRDs while also being able to safely reinstall an operator when needed without loosing and CRs.

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>