ACR replication management revamp (#2366)
* ACR replication management revamp

apparently ACR replication generation can fail during creation and once that happened, the replica needs to be deleted and recreated. this is hard to do with bicep. so this PR replaces the bicep code with a shell step, that does the following

* check if a failed replica exists for an ACR, if yes, delete it
* creates a replica, if non exists (but makes sure to not create a replica in the home region of the ACR)

since deletion and recreation happens in one step, we can leverage the EV2 retry functionality in case the step fails. this will save previous time during region buildout.

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>

* support dry-run

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>

---------

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>