denyassignment env variable for CS
this PR introduces a new config options `clustersService.denyassignments` which controls if the denyassignment feature is enabled in a specific environment.
we will enable this for the time being in STAGE while the feature is still under development. for any other environment this feature will remain disabled.

the override to enable denyassignments can be found in the sdp-pipelines repo the the MSFT sensitive configuration overrides

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>