api: Disallow "openshift-v" prefix on version IDs
The custom "openshift_version" validation tag is similar to the
baked-in "semver" tag, except it allows semantic version values
to have an optional "openshift-v" prefix.

But now that the frontend strips this prefix off when returning
a semantic version ID in a response body, the frontend will no
longer accept this prefix in a request body.

So we now use the baked-in "semver" validation tag for semantic
version ID fields in cluster and node pool resources.

https://pkg.go.dev/github.com/go-playground/validator/v10#hdr-Semantic_Version