tooling/templatize: honor dependencies during step execution
Previously, we were doing a naive in-order traversal of the resource
groups and steps, which led some piplines to be written in very
convoluted ways so as to make sure that this implementation detail did
not interfere with execution and input/output chaining.

We now build a graph of steps that need to execute, and traverse them in
a depth-first manner. We will be able to add parallelism to this as
well, but I did not have time to make that happen right now.

Signed-off-by: Steve Kuznetsov <stekuznetsov@microsoft.com>