tooling/templatize: signal consumers to exit with context
Previously, a consumer routine would close the output channels for all
consumers before exiting, but this would cause other consumers to
attempt to send to a closed channel if they exited later. Instead, we
can use a scoped context specifically for consumers to signal to other
consumers that we encountered an error and need to close. Waiting for
all consumers to exit before closing the channels ensures we don't panic.

Signed-off-by: Steve Kuznetsov <stekuznetsov@microsoft.com>