tooling/templatize: use require over assert
The assert library does the equivalent of `t.Error()`, whereas in all of
the cases I reviewed in this test, we want `t.Fatal()`. If we must use a
special library to save us one `if` statement (:eyeroll:) we can use the
`require` library to get the behavior we want.

Signed-off-by: Steve Kuznetsov <stekuznetsov@microsoft.com>