api: Drop "required_for_put" validation tag
This is a :facepalm: commit.

I don't know why I had it in my head that required field validation
had to be restricted to PUT requests. Maybe it was true at one time
during early development and got lodged in my brain.

My concern at the time was with PATCH requests. But validation runs
on the struct of an existing resource *after* the PATCH request body
is applied, not on the PATCH request body itself. The patched struct
must still have non-empty values for all required fields.