api: Managed service identity is technically optional
This is purely for correctness.

Even though the hcpOpenShiftCluster resource is configured to use
user-assigned managed identities in the ARM manifest, it still is
technically an optional field and should be represented as such in
the API model.

Additionally, managed service identity type is a *required* field
and does not have a default value.