frontend: Update OCM tests
Test with default resource instances instead of empty instances.
Empty instances may not be valid (e.g. empty enum fields).

Also, rework how to construct OCM objects for test cases. The
builder API is not intended for multiple passes, but test cases
are structured to take a default object and tweak it.

How this works now is test cases can still define object builders,
but each builder is marshaled to a JSON byte array, unmarshaled to
a generic map (map[string]interface{}) and merged with an existing
generic map.

Then finally we take the generic map with all the merged data,
marshal it to a JSON byte array, and unmarshal it back to an OCM
object.