reorder global pipeline steps
we want to depend on output chaining based on the output step of the global pipeline, not the infra step.

this way we can skip the infra step if it gives us troubles and all followup steps can still complete (under certain conditions, e.g. infra run at least once correctly to completion)

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>