*: add MSFT IcM alert group, send some metrics there
For now, these alerts are copy-pasta from the broader group of
generated alerts, as the tooling in this repo does not have the capacity
to send some but not all generated alerts to a specific alert group. We
are OK taking on the tech debt of copy-pasting these, for now, until
there is time to fixup the generation tooling.

Signed-off-by: Steve Kuznetsov <stekuznetsov@microsoft.com>