Add Kiota SDK Generation: MSGraph
The Microsoft Graph SDK is massive, this PR introduces an alternative to
the full SDK: generating our own from the openapi spec using Kiota. This
PR gives us a path forward on migrating out of
microsoftgraph/msgraph-sdk-go and into ARO-HCP/internal/graph/graphsdk,
which gives us the following savings:

- 191MB SDK -> 27MB SDK
- 10,000+ endpoints -> 7 endpoints
- Faster builds, smaller binaries, reduced security risk

The solution also allows us to follow this pattern for other SDKs in the
future as needed, it'll take a few lines of Bash to get there.

The actual MSGraph SDK will be in a follow-up commit.