api: Halt validation on field visibility errors
The rationale is if a field cannot be changed in the first place,
then it doesn't matter if the new value for the field is valid.
There's no need for two validation errors on the same field.

So there's now three phases of static validation:

1) Field visibility
2) Single-field validation based on struct tags
3) Multi-field validation through custom routines

Errors in any phase will abort subsequent phases.