api: Add VisibilityNullable flag
VisibilityNullable is a computed visibility flag that is set
when the internal struct field has a pointer type and has the
VisibilityUpdate flag. It means the field can be set to "null"
in a PATCH request.

Attempting to set fields without this flag to "null" in a PATCH
request will result in an error during visibility validation.

For example, patching an hcpOpenShiftClusters resource with:

  {
    "properties": null
  }

causes a validation error response:

  {
    error: {
      code: "InvalidRequestContent",
      message: "Field 'properties' cannot be removed",
      target: "properties"
    }
  }