frontend: Refactor ArmResourceList
Resources of type hcpOpenShiftVersions are not stored in Cosmos DB.
Avoid an unnecessary database query when listing this resource type.

It does mean pagination will not work correctly for this resource
type since it requires a Cosmos DB continuation token, but I'm not
anticipating we'll have enough instances to require pagination.