api: Override CIDR field names in generated clients
I don't remember if Azure mandates this, but the spelling of the
acronym CIDR in our TypeSpec models is unfortunate (e.g. "PodCidr"
instead of "PodCIDR").

It's probably too late to correct this in the ARM API, but we can
still do so in generated clients. It turns out there's a TypeSpec
decorator "@clientName" for overriding generated field names.