tooling/prometheus-rules: allow udpating test fixtures
It is not clear to me why these tests were not written to use the normal
golden fixture approach we have, or why they are all factored to use
implicit data passing using the filesystem. A better refactor would test
a function that takes config and data as input (as opposed to references
to files) and asserts over the output, not over output files. That is a
larger refactor; keeping this surgical, I added the smallest diff to get
us useful updates to test fixtures.

Signed-off-by: Steve Kuznetsov <stekuznetsov@microsoft.com>