tooling/prometheus-rules: expose title, correlation ID to IcM
We want the IcM titles to be more meaningful - right now they contain a
bunch of useless or semi-useless information. By explicitly setting the
field, we can use the nice templated descriptions and highlight the
cluster.

Similarly, the default correlation ID does not take into account the
dimensions along which the alert varies, which are necessary to include
in the correlation ID to make sure that separate instances of the alert
firing (across different clusters, or for different workloads in one
cluster) get separate IcMs.

Signed-off-by: Steve Kuznetsov <stekuznetsov@microsoft.com>