tooling/templatize: allow using an explicit record of sub ids
The existing lookup logic for subscription IDs is clever - in the
development config, we use the subscription display name as the 'key'
for Ev2, since Ev2 never touches these anyway. However, when trying to
use the local runner to target actual subscriptions known to Ev2 with a
real key, no longer will the display name of the subscription be the
same as the 'key' in our config. We can expose a literal mapping of
subscription key to ID in the dev settings file, though, as this set
changes infrequently and is well-known. Then, the lookup is simple.

Signed-off-by: Steve Kuznetsov <stekuznetsov@microsoft.com>