tooling/prometheus-rules: duplicate description as title
It looks like some part of the Azure Monitor stack very helpfully takes
the `annotations.description` field from a Prometheus rule and slurps it
into some higher-level alert field, thereby removing it entirely from
the alert context data. We want to use this value in our IcM titles,
though, so we need to have it - it is extremely difficult to understand
the data format for the alert and how to refer to the new field that
the monitor puts it into; so, we just duplicate the annotation and we
to refer to the annotation instead.

Signed-off-by: Steve Kuznetsov <stekuznetsov@microsoft.com>