frontend: Return builder objects when converting from RP to CS
The purpose of this is to decouple the conversion functions from
the frontend. The problem has been the Properties field that gets
added to ClusterBuilder objects. The Properties field has various
configuration settings that are held by the ClusterServiceClient,
and the frontend has the ClusterServiceClient.

By returning builder objects, we can delay adding the Properties
field until it's handed off to the ClusterServiceClient instance.
This is only really needed for ClusterBuilder objects; the other
builder functions are being changed only to maintain consistency.