api: Fix visibility validation bug
For the purpose of validation, treat nil slices and empty slices as
equal when comparing values. Same goes for maps.

Treat a nil pointer and a pointer to a zero-value as equal only when
the field is not nullable. (A nullable field treats a nil pointer as
distinct from a pointer to a zero value.)

Also add some unit test cases to reproduce and verify the bug.

This came up because our E2E cluster bicep template specifies

  console: {}

and the "console" field is read-only (and not nullable).

Submitting a read-only field is allowed as long as the field value
matches the resource's current value.

Visibility validation was comparing the incoming &ConsoleProfile{}
value to the resource's internal nil pointer and mistreating it as
an attempt to change a read-only field.