dev-infrastructure: don't use outputs from ARM steps
We want to allow for incremental releases to skip steps when running
them would not yield a different outcome from the last time they
successfully ran. However, we have a couple of places where large ARM
deployment steps output variables that are chained into other steps. If
any of those derivative steps needs to run, the big ARM deployment will
be pulled into the execution graph as well to provide the output
variables. By using a dedicated output-only step instead, we allow for
the large deployment step to be skipped and only require the output
step.

Signed-off-by: Steve Kuznetsov <stekuznetsov@microsoft.com>