tooling: track ill-formed steps, ensure we don't add any more
This test will make people pause before trying to add more ill-formed
steps, and act as a nice tracker for what we have left to do. Once we
reach a state where all of our steps are well-formed, we can remove the
output tracker and assert that no ill-formed steps are added, as opposed
to asserting that the set does not change.

Signed-off-by: Steve Kuznetsov <stekuznetsov@microsoft.com>