rework schema
the current config model and ev2 translation does not allow for unbounded list of config options. hence we went for named channelgroups (stable and fast at this time as CS does only support them).

this is more involved because more values need to be carried from config -> pipeline.yaml -> Makefile -> values.yaml -> configmap template. the pain of this will ease once we move CS over to the native helm step as templatization can happen directly in the values.yaml file and we can skip pipelin.yaml and makefile configuration alltogether.

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>