split the create and update cluster functions
This is a precursor to creating new CRUD functions for these and it
eliminates branching.  Additionally, it allows for not reusing instance
variables for different purposes, so we get more precise names to follow
the code.

This further exposes some manipulation that ends up extraneous when we
start directly storing the internal Cluster type, which is what led to
this refactor.