cloud specific CS operator role sets in helm chart
declare premade sets of operator roles per deployment cloud (`dev` being the RH tenant, `prod` being AME). config.yaml only specifies the roleSetName to be used for a specific environment.

also the lookup of the operator role IDs has been removed. the IDs are now specified along with the role names in the values.yaml file. reasoning: we need to switch to the native helm step soon where such lookups with shell commands are not supported. the usual way would be lookup in bicep and forward the data to the helm chart, but roles can't be looked up by name via bicep yet. there is a fix in the working on bicep side to support that.

due to parameter count limitations we need to specify the operator role names within the helm values file instead of the configuration.

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>