Add a Helm resource policy annotation to avoid secret deletion
This is a required step to split the CS chart in two and move the shard
registration secret to a different chart. Without this annotation, the
CS chart will delete the live resource if the template is moved out of
its control.

Pre-requisite to merge https://github.com/Azure/ARO-HCP/pull/2930. A
release with this change needs to be deployed all the way up to
production to make the split a safe operation.