update ocm-sdk-go to v0.1.479
It also includes an update of the github.com/openshift-online/ocm-api-model
modules containing the ocm API models.

The update of the SDK and ocm API models contain the changes to
split/duplicate the ARO-HCP API models from the OCM commercial ones,
with the exception of the break glass models/endpoints. Those still
point to OCM commercial API models. Once the implementation in CS moves
to offer those endpoints under the aro_hcp base endpoint we will be able
to also move that one too.

This commit also contains changes needed to be able to compile with
the new ocm-sdk-go version where we duplicated some data types to the
corresponding arohcp package.