Fix RMO integration
Stricter validation has been added to RMO and the controller fails if
the ingresscontroller "default" resource is not present in the
openshift-ingress-operator namespace. Furthermore, the controller
expects that the
`.Status.EndpointPublishingStrategy.LoadBalancer.ProviderParameters.AWS`
field is populated. To keep the integration working, the
IngressController CRD as well as the "default" resource are deployed,
with the required status fields present. This makes the controller pass
the added validation and actually reconcile the blackbox-exporter and
the servicemonitor resource we required for KAS monitoring.

This is a hack, a temporary solution until RMO is removed and our own
ARO-HCP specific solution is in place.