*: move Helm script to cluster-service
The only user of this script is CS, and we want to add a working
directory filter on the shell step running their deployment to make it
well formed over inputs. While we work to migrate them to a real Helm
step and delete the shell script, we can move the script into here and
still get incremental releases.

Signed-off-by: Steve Kuznetsov <stekuznetsov@microsoft.com>