*: add Makefile support, templatize env for Prow
When running in Prow, the CI job needs to be able to configure where
timing files and visualizations are stored, as well as using a specific
deployment environment that uses the $BUILD_ID to differentiate the
jobs. This ID is a monotonically increasing integer, globally unique for
the entire Prow instance, not per job, so using the last three digits
should be safe as long as we have fewer than 1000 concurrent jobs.

Signed-off-by: Steve Kuznetsov <stekuznetsov@microsoft.com>