dev-infrastructure: add AKS maintenance windows
Our AKS clusters opt into auto-upgrade for patch-level control plane
upgrades and node image upgrades. While these upgrades are, in the
abstract, fairly safe, they always carry some risk. We do not want to
burden our on-call staff with middle-of-the-night calls or any weekend
calls if we absolutely do not have to, so we need to use maintenance
windows to forbid changes to the cluster during these times. The
maintenance window also provides a simple way to capture CCOA periods.

As proposed, the maintenance window triggers Monday, at 15:00 UTC, and
lasts for 10 hours. This roughly captures the US working timezones for
mountain and west coast time zones.

Signed-off-by: Steve Kuznetsov <stekuznetsov@microsoft.com>