fix: fix hack directory reference in cluster-service/helm.sh
The cluster-service/helm.sh was originally in hack/helm.sh.
Once it was moved to the cluster-service directory the hack
directory reference witihn the helm.sh script broke because
it assumed that the directory of helm.sh was the same as other
scripts used by it. This was making the execution of
the cluster-service/helm.sh script fail, used when deploying
the CS Helm target.

This commit moves the hack/deployment-diagnostics.sh,
hack/helm-adopt.sh and hack/pod-logs.sh scripts into a new directory
cluster-service/hack. We can move them because they are only used by
the cluster-service/helm.sh script. Additionally, we also move
cluster-service/helm.sh into cluster-service/hack/helm.sh and we update
cluster-service/Makefile to change the call to helm.sh to the
new location