frontend: Change response codes when failing to acquire lock
The various Azure SDKs have a fixed set of response codes that by
default trigger a retry:

  408 Request Timeout
  429 Too Many Requests
  500 Internal Server Error
  502 Bad Gateway
  503 Service Unavailable
  504 Gateway Timeout

The response code we've been using for subscription lock contention
(409 Conflict) does not by default trigger a retry in the Azure SDKs,
despite the response including a Retry-After header.

This has been causing Azure deployments via Bicep to fail when they
should be retrying.

Instead return 503 Service Unavailable with a custom Azure error
code indicating lock contention. This could be a clue that the
frontend pods need to be scaled up.