tooling/templaize: correctly handle existing deployments
Our tooling differentiates deployments based on a deterministic hash of
their inputs. If we name our deploymnets in the Azure control plane with
that same hash, we can resume waiting for an existing deployment created
by a previous local run that failed for whatever reason. The existing
code to wait for previous deployments was insufficient, as we were
sharing names across runs when we did not need to, and it would start a
new deployment after waiting for the previous one to succeed.

Signed-off-by: Steve Kuznetsov <stekuznetsov@microsoft.com>