Adjust function naming to reflect cost and return useful type
Until we're able to cleanly transform cosmos API type to internal to
external without secondary lookups, we'll have some data retrieval and
merge functions.  This aligns the data retrieval function to show its
cost and adjusts the return type so we have a type we can re-use the
return value of somewhere internally.  The json.Marshal was ancillary
to the body of the call.