test/Makefile: use actual Make
This commit changes the Makefile in the test directory to use real
`make` rules for the .bicep -> .json translation as well as for the
compilation of the binaries. This allows the main Makefile to import
this one and depend on the test runner for the `e2e-local/run` target,
which in turn allows end-to-end jobs do something like:

make e2e-local/run -o test/aro-hcp-tests

Which will ensure that the Go binary is not rebuilt. We want the Go
binary not to be rebuild (normal `make` rules nonwithstanding) because
"something" is messing with timestamps in the job and causing a spurious
rebuild, and the OCP CI requirement to vendor everything really makes
this repo messy to build.

We should have done the ARM template compilation and Go builds properly
in the first place, they're done right in the rest of the repo.

Signed-off-by: Steve Kuznetsov <stekuznetsov@microsoft.com>