remove unnecessary patches from the cluster create transaction
With the separation of create and update, along with the inclusion of
typed for the database calls, we no longer have a need to perform create
plus patching to create a full view of the object.  it is more clear to
create the resource we want to store and store that.