introduce dedicated kusto subscriptions
introduce a dedicated subscription configuration for kusto. this allows for flexibility in kusto instance creation. depending on the chosen subscription key, a kusto instance can be per geography or even per region if required.

having kusto instances keyed on a subscription allows for things like ARM lookup steps that emit data like the actual kusto instance ingress and read urls, as well as the actual resource id required for cross-subscription permission management in the svc and mgmt pipelines for arobit.

this PR does not touch the implications of adding kusto subscriptions yet. it is merely a way to define out intent and allows us to already create subscriptions in sdp-pipelines, before we need and use them.

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>