avoid using *CloudError for return
in golang, if you assign an instance of the interface to an interface
weird things can happen.  In particular, you can have a non-nil error
with a nil instance of CloudError, which produces strange behavior.
Since all error handling is now coalesed and unwraps, we can avoid
bothering with the distinction in most places.