test: remove invalid E2E scenario
The provisioning state is returned by ARM to indicate the status of the
last long-running operation executed against the hosted cluster (in the
ARM control plane). Not only does this not really have any bearing on
the ability to operate on the node pools in the cluster, but we
explicitly want users to continue being able to edit node pools if, say,
a cluster failed to upgrade correctly.

Signed-off-by: Steve Kuznetsov <stekuznetsov@microsoft.com>