add data dumps just prior to deletion and for every long running operation
We will dump about every five minutes, limited by an LRU cache to
prevent explosions in logging.  Once we have a better way to hit the
admin API, we can remove the periodic dump.  This will grow in value,
the more information we place in cosmos and can be harvested in kusto
and must-gather.