Add command to produce custom-link-tools.html to show on spyglass
Spyglass shows files name custom-link-tools.html on the job run page.
This adds a command to look at step.yaml to determine the service
cluster name we provisioned and uses that to build a dataexplorer URL
that drops you into the kusto view with most of the data already
populated to look at backend or frontend logs.  Others could be added.

We could also analyze the junit to locate failed resource groups and
eventually use our tagging to build a URL for failed tests.