Making deletion idempotent is tempting but it actually causes a problem with Azure's async protocol.
From Matt Barnes:
If we just return nil here, the 202 Accepted response code will indicate an asynchronous operation is underway but the response will be missing the necessary headers Location and Azure-AsyncOperation to indicate the tracking endpoints for the operation.

Simply adding the headers won't work either because the requesting client may not be authorized to read the endpoints already established for this operation.

So we actually need to let checkForProvisioningStateConflict return a 409 Conflict here.

The one exception we could make here is if the same client that started the deletion — determined by comparing the previously created operation record's client and tenant ID to the request headers — re-requests deletion, then and only then could we treat it idempotently. But we'd still have to include those aforementioned response headers. This has been on my "nice to have" list for awhile but never got round to it.