Remove log diagnostics
This was only ever useful in shared dev. Since we have efforts to move to e2e parallel, remove this settings. They are only an additional piece that can break, and history of jobs showed they fail sometimes. In case an engineer this, simply create a workspace and configure it manually as needed