rekey all cosmos data from .id=UUID to .id=resourceID
This change performs a data migration on read to switch from reading
based on a cosmos search of fields matching a pattern to direct item
reads based on key.  On a failure to lookup by cosmosID and success
reading based on existing search, we create a new item and delete the
original.

Since old frontends use the cosmosUID as fully opaque, the new ID
doesn't cause problems for old frontends even on rollback.

When the frontend starts, we find all subscriptions, then search for all
resources in each subscription using the untyped client listRecursive,
then read each item to trigger the migration.